+++ HISTORY.md
@@ -4,8 +4,9 @@
 - [x] rewrite test server in c, test cases in golang
 - [x] update travis ci with new test env
 - [x] implement timeout.set, timeout.get functions
-- [x] implement json_get_double, json_set_double
+- [x] implement json_get_double, json_set_double, json_get_long
 - [x] refactor tid data type
+
 
 ---